AK-30659 Rename connector_cisco_ftdv to service_cisco_ftdv
Rename connector_cisco_ftdv to service_cisco_ftdv.

+ Pick up new alkira-client-go with new function names.
+ Relocate username and password into firepower_management_center block.
+ Rename fmc_username to username to be consistent with portal.
+ Rename fmc_password to password to be consistent with portal.
+ Rename fmc_ip to server_ip to be consistent with portal.
+ Relocate credential into management center.
+ Relocate ip_allow_list into management center.
