+++ alkira/provider.go
@@ -42,7 +42,6 @@ func Provider() *schema.Provider {
 			"alkira_connector_azure_vnet":         resourceAlkiraConnectorAzureVnet(),
 			"alkira_connector_azure_expressroute": resourceAlkiraConnectorAzureExpressRoute(),
 			"alkira_connector_cisco_sdwan":        resourceAlkiraConnectorCiscoSdwan(),
-			"alkira_connector_cisco_ftdv":         resourceAlkiraConnectorCiscoFTDv(),
 			"alkira_connector_gcp_vpc":            resourceAlkiraConnectorGcpVpc(),
 			"alkira_connector_oci_vcn":            resourceAlkiraConnectorOciVcn(),
 			"alkira_connector_internet_exit":      resourceAlkiraConnectorInternetExit(),
@@ -70,6 +69,7 @@ func Provider() *schema.Provider {
 			"alkira_segment_resource":             resourceAlkiraSegmentResource(),
 			"alkira_segment_resource_share":       resourceAlkiraSegmentResourceShare(),
 			"alkira_service_checkpoint":           resourceAlkiraCheckpoint(),
+			"alkira_service_cisco_ftdv":           resourceAlkiraServiceCiscoFTDv(),
 			"alkira_service_fortinet":             resourceAlkiraServiceFortinet(),
 			"alkira_service_infoblox":             resourceAlkiraInfoblox(),
 			"alkira_service_zscaler":              resourceAlkiraServiceZscaler(),
