--- docs/resources/connector_cisco_ftdv.md
@@ -1,130 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "alkira_connector_cisco_ftdv Resource - terraform-provider-alkira"
-subcategory: ""
-description: |-
-  Manage Cisco FTDv Connector. (BETA)
----
-
-# alkira_connector_cisco_ftdv (Resource)
-
-Manage Cisco FTDv Connector. (**BETA**)
-
-## Example Usage
-
-```terraform
-resource "alkira_connector_cisco_ftdv" "cisco_ftdv" {
-  name                = "FtdvFirewalll"
-  auto_scale          = "OFF"
-  size                = "SMALL"
-  tunnel_protocol     = "IPSEC"
-  cxp                 = "US-WEST"
-  max_instance_count  = 1
-  min_instance_count  = 1
-  ip_allow_list       = ["192.168.3.3"]
-
-  global_cidr_list_id = alkira_list_global_cidr.checkpoint_cidr.id
-  segment_ids = [alkira_segment.tftest1.id, alkira_segment.tftest2.id]
-
-  username = "user1"
-  password = "Abcd1234567"
-
-  management_server {
-    fmc_ip       = "1.1.1.1"
-    segment_id   = alkira_segment.tftest1.id
-  }
-
-  instances {
-    admin_password       = "Abcd@12345"
-    fmc_registration_key = "abcd12345"
-    ftdv_nat_id          = "abcd1234"
-    version              = "7.2.1-40"
-    license_type         = "BRING_YOUR_OWN"
-  }
-
-  segment_options {
-    segment_id = alkira_segment.tftest1.id
-    zone_name  = "zone1"
-  }
-  segment_options {
-    segment_id = alkira_segment.tftest2.id
-    zone_name  = "zone2"
-  }
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `cxp` (String) The CXP where the connector should be provisioned.
-- `global_cidr_list_id` (Number) The ID of the global cidr list to be associated with the management server. The global cidr list must be tagged with `CISCO FTDV`.CIDR must be at least /25.
-- `instances` (Block Set, Min: 1) (see [below for nested schema](#nestedblock--instances))
-- `management_server` (Block Set, Min: 1) The Firepower Management Server options. (see [below for nested schema](#nestedblock--management_server))
-- `max_instance_count` (Number) The maximum number of Cisco FTDv instances that should be deployed.
-- `name` (String) The name of the connector.
-- `password` (String) Firepower Management Server password.
-- `segment_ids` (List of String) IDs of segments associated with the service.
-- `size` (String) The size of the connector, one of `SMALL`, `MEDIUM`, `LARGE`, `2LARGE`.
-- `username` (String) Firepower Management Server username.
-
-### Optional
-
-- `auto_scale` (String) Indicate if `auto_scale` should be enabled for your Cisco FTDv connector. `ON` and `OFF` are accepted values. `OFF` is the default if field is omitted.
-- `billing_tag_ids` (List of Number) IDs of Billing Tags.
-- `ip_allow_list` (List of String) List of IP Addresses and CIDRs to access the Firepower Management Center.
-- `min_instance_count` (Number) The minimum number of Cisco FTDv instances that should be deployed.
-- `segment_options` (Block Set) The segment options used by the Cisco FTDv. (see [below for nested schema](#nestedblock--segment_options))
-- `tunnel_protocol` (String) The tunnel protocol. Default is `IPSEC`
-
-### Read-Only
-
-- `credential_id` (String) An opaque identifier generated when storing Cisco FTDv credentials.
-- `id` (String) The ID of this resource.
-
-<a id="nestedblock--instances"></a>
-### Nested Schema for `instances`
-
-Required:
-
-- `admin_password` (String) Cisco Firepower Firewall instance admin password.
-- `fmc_registration_key` (String) Cisco Firepower Firewall instance registration key.
-- `license_type` (String) Cisco Firepower Firewall instance license type, either `BRING_YOUR_OWN` or `PAY_AS_YOU_GO`.
-- `version` (String) The version of the Cisco Firepower Firewall instance.
-
-Optional:
-
-- `ftdv_nat_id` (String) ID of NAT which FTDv Services sit behind.
-- `hostname` (String) Name of the Cisco Firepower Firewall instance. If empty will use Cisco FTDv `name` field.
-
-Read-Only:
-
-- `credential_id` (String) An opaque identifier generated when storing Cisco Firepower Firewall instance credentials.
-- `id` (Number) ID of the Cisco Firepower Firewall instance.
-- `internal_name` (String) Generated internal name when storing Cisco Firepower Firewall instance.
-- `state` (String) Internal state of the Cisco Firepower Firewall instance.
-
-
-<a id="nestedblock--management_server"></a>
-### Nested Schema for `management_server`
-
-Required:
-
-- `fmc_ip` (String) IP address of the Firepower Management Server.
-- `segment_id` (Number) ID of the segment accociated with the Firepower Management Server.
-
-
-<a id="nestedblock--segment_options"></a>
-### Nested Schema for `segment_options`
-
-Required:
-
-- `segment_id` (Number) ID of the segment.
-- `zone_name` (String) The name of the associated zone.
-
-Optional:
-
-- `groups` (List of String) The list of Groups associated with the zone.
-
-
