+++ docs/resources/service_cisco_ftdv.md
@@ -1,0 +1,134 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "alkira_service_cisco_ftdv Resource - terraform-provider-alkira"
+subcategory: ""
+description: |-
+  Manage Cisco FTDv Service. (BETA)
+---
+
+# alkira_service_cisco_ftdv (Resource)
+
+Manage Cisco FTDv Service. (**BETA**)
+
+## Example Usage
+
+```terraform
+resource "alkira_service_cisco_ftdv" "test" {
+  name               = "test"
+  auto_scale         = "OFF"
+  size               = "SMALL"
+  tunnel_protocol    = "IPSEC"
+  cxp                = "US-WEST"
+  max_instance_count = 1
+  min_instance_count = 1
+
+  global_cidr_list_id = alkira_list_global_cidr.test.id
+  segment_ids         = [alkira_segment.test1.id, alkira_segment.test2.id]
+
+  firepower_management_center {
+    username      = "user1"
+    password      = "Abcd1234567"
+    server_ip     = "1.1.1.1"
+    segment_id    = alkira_segment.test1.id
+    ip_allow_list = ["192.168.3.3"]
+  }
+
+  instance {
+    hostname             = "instance1"
+    admin_password       = "Abcd@12345"
+    fmc_registration_key = "abcd12345"
+    ftdv_nat_id          = "abcd1234"
+    version              = "7.2.1-40"
+    license_type         = "BRING_YOUR_OWN"
+  }
+
+  segment_options {
+    segment_id = alkira_segment.test1.id
+    zone_name  = "zone1"
+  }
+  segment_options {
+    segment_id = alkira_segment.test2.id
+    zone_name  = "zone2"
+  }
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `cxp` (String) The CXP where the service should be provisioned.
+- `firepower_management_center` (Block Set, Min: 1) The Firepower Management Center options. (see [below for nested schema](#nestedblock--firepower_management_center))
+- `global_cidr_list_id` (Number) The ID of the `alkira_list_global_cidr` to be associated with the service. The list must be tagged with `CISCO FTDV`. CIDR must be at least `/25`.
+- `instance` (Block List, Min: 1) (see [below for nested schema](#nestedblock--instance))
+- `max_instance_count` (Number) The maximum number of instances that should be deployed.
+- `name` (String) The name of the service.
+- `segment_ids` (List of String) IDs of segments associated with the service.
+- `size` (String) The size of the service, one of `SMALL`, `MEDIUM`, `LARGE`, `2LARGE`.
+
+### Optional
+
+- `auto_scale` (String) Indicate if `auto_scale` should be enabled for your Cisco FTDv service. `ON` and `OFF` are accepted values. Default is `OFF`.
+- `billing_tag_ids` (List of Number) IDs of Billing Tags.
+- `min_instance_count` (Number) The minimum number of instances that should be deployed.
+- `segment_options` (Block Set) The segment options used by the Cisco FTDv. (see [below for nested schema](#nestedblock--segment_options))
+- `tunnel_protocol` (String) The tunnel protocol. Default is `IPSEC`.
+
+### Read-Only
+
+- `id` (String) The ID of this resource.
+
+<a id="nestedblock--firepower_management_center"></a>
+### Nested Schema for `firepower_management_center`
+
+Required:
+
+- `password` (String) Firepower Management Center (FMC) password.
+- `segment_id` (Number) ID of the segment accociated with the Firepower Management Center.
+- `server_ip` (String) IP address of the Firepower Management Center.
+- `username` (String) Firepower Management Center (FMC) username.
+
+Optional:
+
+- `ip_allow_list` (List of String) List of IP addresses and CIDRs to access the Firepower Management Center.
+
+Read-Only:
+
+- `credential_id` (String) An opaque identifier generated when storing firepower_management_center credentials.
+
+
+<a id="nestedblock--instance"></a>
+### Nested Schema for `instance`
+
+Required:
+
+- `admin_password` (String) Firepower Firewall Admin Password.
+- `fmc_registration_key` (String) FMC Registration Key.
+- `hostname` (String) Hostname of the Firepower Firewall.
+- `license_type` (String) Cisco Firepower Firewall license type, either `BRING_YOUR_OWN` or `PAY_AS_YOU_GO`.
+- `version` (String) Cisco Firepower Firewall version.
+
+Optional:
+
+- `ftdv_nat_id` (String) FTDv NAT ID.
+
+Read-Only:
+
+- `credential_id` (String) An opaque identifier generated when storing Cisco Firepower Firewall instance credentials.
+- `id` (Number) ID of the Cisco Firepower Firewall instance.
+
+
+<a id="nestedblock--segment_options"></a>
+### Nested Schema for `segment_options`
+
+Required:
+
+- `segment_id` (Number) ID of the segment.
+- `zone_name` (String) The name of the associated zone.
+
+Optional:
+
+- `groups` (List of String) The list of Groups associated with the zone.
+
+
